First version of The Flood
Modifies factorio so that a flood creeps across the map, destroying everything in its path.
The flood moves from left to right, moving on tile every 600 ticks (10 seconds).
The code checks new chunks when they are generated, and updates existing ones when the flood moves.
It also checks the spawn locations for every force, and moves it so that it is in front of the flood as needed.